Refactor: 메인 화면 수정
